+++ include/uhttp/net/Socket.h
@@ -39,11 +39,11 @@ public:
 
 	bool connect(const std::string &addr, int port);
 
-	ssize_t send(const char *cmd, ssize_t cmdLen);
+	ssize_t send(const char *cmd, size_t cmdLen);
 	ssize_t send(const std::string &cmd);
 	ssize_t send(const char c);
 
-	ssize_t recv(char *buffer, int bufferLen);
+	ssize_t recv(char *buffer, size_t bufferLen);
 
 private:
